requirements.txt
aiogram
openai
requests
beautifulsoup4
lxml
numpy
python-dotenv
fastapi
uvicorn
PyPDF2
